+++ CHANGELOG.md
@@ -6,7 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 <!-- ====================================================================== -->
-## [Unreleased]
+<!-- ## [Unreleased] -->
+
+<!-- ====================================================================== -->
+## [0.5.0] - 2019-08-26
 
 ### Added
 - `Nodaire::Indental` and `Nodaire::Tablatal` now implement `Enumerable`, which
@@ -74,7 +77,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   returning an array of hashes or a CSV string.
 
 <!-- ====================================================================== -->
-[Unreleased]: https://github.com/slisne/nodaire/compare/v0.4.0...HEAD
+[Unreleased]: https://github.com/slisne/nodaire/compare/v0.5.0...HEAD
+[0.5.0]: https://github.com/slisne/nodaire/releases/tag/v0.5.0
 [0.4.0]: https://github.com/slisne/nodaire/releases/tag/v0.4.0
 [0.3.0]: https://github.com/slisne/nodaire/releases/tag/v0.3.0
 [0.2.0]: https://github.com/slisne/nodaire/releases/tag/v0.2.0